Delete unused Electron IPC related functionality in .NET app
The ability to push messages using Electron's IpcMain is not used. Since we're using SignalR as the main IPC method to push messages to clients on both Electron and Browser platforms there's little benefit keeping this around.

Electron IpcMain can be more performant. But so far I have not encountered a bottleneck to justify maintaining 2 IPC implementations.

Removing this and using a single implementation that works in any environment will help if we move away from Electron.